Tested "cube inside a cube" to test the peering into the toilet bowl portal mechanic.
Turns out VRTK does an automatic fade to black if you look inside an opaque surface. But if you put a visible object within the opaque object, after the fade to black, it fades the newly visible object. This is awesome! Really desirable free behavior for my project needs.

The rest of the world does render when the new object fades in, but I believe I can block that by putting opaque materials inside of the object we are peering into, facing back toward the player.