Got it to the point where I can press a button and make the seat fall.
Seems to take greater force in VR vs editor... not sure why.
I had to write a script that enables gravity and adds a little push to get it to fall. Then I have to disable gravity when it gets a certain way closed, or it will fall too fast and go past the limit of the joint. Also, if you have to disable gravity on the rigidbody again, or it will become jittery.