Flush lever can move appropriately with player interaction
This was a pain in the ass. When the drive is running, it's too powerful for the player to use the lever. So I have to turn it off and on using the grab / release events.

Then I was also getting horrible jittering while the drive tried to hit the target value. I had to find and increase the target threshold, which was buried somewhere in the VRTK prefab.